app/backend/requirements.txt
alembic==1.13.2
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.11.0
Babel==2.13.1
bcrypt==4.1.1
bidict==0.23.1
billiard==4.2.0
blinker==1.7.0
blis==1.3.3
cachelib==0.9.0
cachetools==6.2.2
catalogue==2.0.10
celery==5.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.13
dnspython==2.4.2
docx2txt==0.9
email-validator==2.1.0.post1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
eventlet==0.40.3
fastapi==0.111.0
fastapi-cli==0.0.5
Flask==2.2.5
Flask-BabelEx==0.9.4
Flask-Caching==2.3.0
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security==3.0.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-speech==2.34.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.0.1
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
idna==3.6
importlib_resources==6.5.2
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.5.2
kombu==5.4.0
Mako==1.3.5
markdown-it-py==4.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
murmurhash==1.0.15
nltk==3.9.2
numpy==2.3.4
orjson==3.11.4
packaging==24.2
passlib==1.7.4
pillow==12.0.0
preshed==3.0.12
prompt_toolkit==3.0.47
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.8.0
pyparsing==3.2.5
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.12.3
python-multipart==0.0.20
python-socketio==5.15.0
pytz==2023.3.post1
PyYAML==6.0.3
redis==5.0.8
regex==2025.11.3
reportlab==4.4.4
requests==2.32.5
rich==14.2.0
rsa==4.9.1
setuptools==71.1.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
smart_open==7.5.0
sniffio==1.3.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speaklater==1.3
SQLAlchemy==2.0.23
srsly==2.5.2
starlette==0.37.2
thinc==8.3.10
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.1
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.29.0
uvloop==0.22.1
vine==5.1.0
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.2.13
weasel==0.4.3
websockets==15.0.1
Werkzeug==2.2.2
wrapt==2.0.1
wsproto==1.3.2
WTForms==3.1.1
azure-cognitiveservices-speech
